Stop menu re-rendering on change of survey screen
