+++ packages/datatrak-web/src/features/Survey/index.ts
@@ -6,5 +6,6 @@
 export * from './Screens';
 export { SurveyContext, useSurveyForm, getArithmeticDisplayAnswer } from './SurveyContext';
 export { SurveyLayout } from './SurveyLayout';
-export { SurveyToolbar, CancelSurveyModal } from './Components';
+export { SurveyToolbar, CancelSurveyModal, SurveySideMenu } from './Components';
 export { getAllSurveyComponents } from './utils';
+export { useValidationResolver } from './useValidationResolver';
