+++ packages/datatrak-web/src/features/index.ts
@@ -14,6 +14,8 @@ export {
   useSurveyForm,
   getAllSurveyComponents,
   CancelSurveyModal,
+  SurveySideMenu,
+  useValidationResolver,
 } from './Survey';
 export { RequestProjectAccess } from './RequestProjectAccess';
 export { MobileAppPrompt } from './MobileAppPrompt';
